refac(fe):editor head component separation
