+++ frontend/nevvel/src/types/editor/index.d.ts
@@ -60,4 +60,12 @@ declare module "editor" {
     url: string;
     uploader: AssetUploader;
   }
+  
+  export interface date {
+    year: number;
+      month: number;
+      date: number;
+      hours: number;
+      minutes: string;
+  }
 }
